--- com.unity.renderstreaming/Runtime/Scripts/RenderPipeline.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d3a3fc772d2874c27a543c6a5f466ada
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
