--- com.unity.renderstreaming/Runtime/Scripts/RenderPipeline/LegacyRPRenderTextureBlitter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 672c5793bed422e42a82e06e24fc4a0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
