--- com.unity.renderstreaming/Runtime/Scripts/RenderPipeline/URPRenderTextureBlitter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 09dd73f6bc80ce84b954b9619472979a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
