+++ com.unity.renderstreaming/Samples~/Scenes/SimpleRenderStreamingLegacyRP.unity
@@ -1197,7 +1197,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 963194225}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 672c5793bed422e42a82e06e24fc4a0a, type: 3}
+  m_Script: {fileID: 11500000, guid: 88167b61ef2361446aefab18c3c7a843, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_rtCamera: {fileID: 725451154}
