+++ com.unity.renderstreaming/Samples~/Scenes/SimpleRenderStreamingURP.unity
@@ -1364,7 +1364,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 963194225}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 09dd73f6bc80ce84b954b9619472979a, type: 3}
+  m_Script: {fileID: 11500000, guid: 88167b61ef2361446aefab18c3c7a843, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_rtCamera: {fileID: 725451154}
